app/requirements.txt
Django==2.1.2
gunicorn==19.9.0
psycopg2-binary==2.7.5
djangorestframework==3.8.2
djangorestframework-gis==0.14

erddapy==0.2.0
xarray==0.11.0
netCDF4==1.4.2

sentry-sdk==0.5.5

# dev
ipython==7.0.1
django-debug-toolbar==1.11